safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-stdDeviation-prop-expected.txt

CONSOLE MESSAGE: line 38: TypeError: 'undefined' is not a function (evaluating 'dropShadowElement.setStdDeviation(10, 10)')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stdDeviation' property of the SVGFEDropShadowElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL dropShadowElement.stdDeviationX.baseVal should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'dropShadowElement.stdDeviationX.baseVal')
FAIL dropShadowElement.stdDeviationY.baseVal should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'dropShadowElement.stdDeviationY.baseVal')
